pyproject.toml
[virtualenvs]
in-project = true

[tool.pyright]
exclude = ["**/__pycache__"]
venvPath = "~/.cache/pypoetry/virtualenvs"
venv = "kp-WfgNVHx5-py3.10"

[tool.poetry]
name = "kp"
version = "0.1.0"
description = ""
authors = ["super <stijnvanderpouw@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
alembic = "1.11.1"
anyio = "3.6.2"
asyncpg = "0.27.0"
bcrypt = "4.0.1"
cffi = "1.15.1"
click = "8.1.3"
cryptography = "40.0.2"
ecdsa = "0.18.0"
fastapi = "0.95.2"
greenlet = "2.0.2"
h11 = "0.14.0"
httptools = "0.5.0"
idna = "3.4"
mako = "1.2.4"
markupsafe = "2.1.2"
passlib = "1.7.4"
psycopg2-binary = "2.9.6"
pyasn1 = "0.5.0"
pycparser = "2.21"
python-dotenv = "1.0.0"
python-jose = "3.3.0"
pyyaml = "6.0"
rsa = "4.9"
six = "1.16.0"
sniffio = "1.3.0"
starlette = "0.27.0"
typing-extensions = "4.6.2"
uvicorn = "0.22.0"
watchfiles = "0.19.0"
websockets = "11.0.3"
python-multipart = "^0.0.6"
sqlalchemy = "^2.0.15"
pydantic = {extras = ["email"], version = "^1.10.8"}
email-validator = "^2.0.0.post2"
requests = "^2.31.0"
msgspec = "^0.15.1"
pytest = "^7.4.0"
httpx = "^0.24.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
